07-26/requirements.txt
pika<2.0.0
pycryptodome<4.0.0